ported config
